feat : Various feature improvements  - applying i3dm geoTiff elevation values,  - separated the lod debug option feature,  - change pbrMetallicRoughness value from 0.0 -> 0.5

+++ core/src/main/java/com/gaia3d/basic/exchangable/GaiaSet.java
@@ -111,7 +111,11 @@ public class GaiaSet implements Serializable{
                     imageTempPath.toFile().mkdir();
 
                     Path outputPath = imageTempPath.resolve(diffusePath);
-                    FileUtils.copyFile(new File(imagePath), outputPath.toFile());
+
+                    File outputPathFile = outputPath.toFile();
+                    if (!outputPathFile.exists()) {
+                        FileUtils.copyFile(new File(imagePath), outputPath.toFile());
+                    }
                 }
                 material.write(stream);
             }
